requirements_sync.txt
supabase>=2.3.0
openai>=1.12.0
schedule>=1.2.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
